Move StepBundlesConfig under unified-editor
